override_package_domain: Move package name deduction to a common util package
It is needed beyond just service_factory. This also necessitated removing a
duplicate implementation of snake_to_upper_camel.
